+++ src/platform.test.ts
@@ -98,7 +98,7 @@ describe('ShellyPlatform', () => {
       matterbridgeDirectory: './jest/matterbridge',
       matterbridgePluginDirectory: './jest/plugins',
       systemInformation: { ipv4Address: undefined, osRelease: 'xx.xx.xx.xx.xx.xx', nodeVersion: '22.1.10' },
-      matterbridgeVersion: '2.2.1',
+      matterbridgeVersion: '2.2.4',
       edge: false,
       log: mockLog,
       getDevices: jest.fn(() => {
@@ -220,7 +220,7 @@ describe('ShellyPlatform', () => {
   it('should throw because of version', () => {
     mockMatterbridge.matterbridgeVersion = '1.5.4';
     expect(() => new ShellyPlatform(mockMatterbridge, mockLog, mockConfig as any)).toThrow();
-    mockMatterbridge.matterbridgeVersion = '2.2.1';
+    mockMatterbridge.matterbridgeVersion = '2.2.4';
   });
 
   it('should call onStart with reason and start mDNS', async () => {